src/test/java/com/leetcode/dp/medium/TargetSum_494Test.java
package com.leetcode.dp.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class TargetSum_494Test {

  private final TargetSum_494 solution = new TargetSum_494();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments(new int[] {1, 1, 1, 1, 1}, 3, 5), arguments(new int[] {1, 1}, 2, 1));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void findTargetSumWays(int[] input, int S, int expectedResult) {
    int actualResult = solution.findTargetSumWays(input, S);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
